# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareBank Backend",
    "description": "Personalized Banking & Financial Wellness API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/auto-savings/recommend": {
      "get": {
        "tags": [
          "auto-savings"
        ],
        "summary": "Get Recommendation",
        "description": "Get auto-savings recommendation based on current financial forecast.",
        "operationId": "get_recommendation_api_auto_savings_recommend_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsRecommendation"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auto-savings/approve": {
      "post": {
        "tags": [
          "auto-savings"
        ],
        "summary": "Approve Savings",
        "description": "Approve and execute a micro-savings transfer.",
        "operationId": "approve_savings_api_auto_savings_approve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveRequest": {
        "properties": {
          "amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "Amount to transfer to savings"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "ApproveRequest"
      },
      "ApproveResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ApproveResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SavingsRecommendation": {
        "properties": {
          "suggested_amount": {
            "type": "integer",
            "title": "Suggested Amount",
            "default": 0
          },
          "goal_progress": {
            "type": "number",
            "title": "Goal Progress",
            "default": 0.0
          },
          "safety_threshold": {
            "type": "number",
            "title": "Safety Threshold",
            "default": 0.0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "title": "SavingsRecommendation"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}